Update node to 16.16 & run `npm update`
Due to an update in the default repositories in the underlying alpine image, python no longer refers to the correct version for our build. Changing it to python3 will allow the build to complete succesfully. See https://stackoverflow.com/questions/62169568/docker-alpine-linux-python-missing

Ideally we'd update to version 18 straight away, but that seems a bit less straightforward. Updating to version 16 (which has security support until september 11th 2023) buys us time to come up with a good strategy to upgrade further.